requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
tavily-python>=0.2.0
python-dotenv>=1.0.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
pydantic>=2.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
transformers>=4.36.0
torch>=2.1.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
python-docx>=1.0.0
lxml>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-magic>=0.4.27
pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.0.0
numpy>=1.24.0
aiofiles>=23.1.0
python-slugify>=8.0.1 